Add classic and dynamic graphs state storing for SSSP and Fix bug with ConCompAlg with unchanged double type in dynamic method

+++ src/main/scala/org/ilmenau/groupstudy/flinkdynamicgraph/algorithms/ConnectedComponentsAlgorithm.scala
@@ -30,7 +30,7 @@ object ConnectedComponentsAlgorithm {
     result
   }
 
-  def runDynamic(graph: Graph[Integer, Double, Integer], addedEdges: Seq[Edge[Integer, Integer]]): Seq[(Integer, Integer)] = {
+  def runDynamic(graph: Graph[Integer, Integer, Integer], addedEdges: Seq[Edge[Integer, Integer]]): Seq[(Integer, Integer)] = {
     if (_disjointSet == null) {
       _disjointSet = new DisjointDataSet[Integer](mutable.HashMap.empty[Integer, Integer])
       graph.getVertices.collect().foreach(e => _disjointSet.makeSet(e.getId))
